requirements.txt
tqdm
requests
psycopg2-binary
pandas
sqlalchemy
jupyterlab
